refactor: decouple core from zod behind a validator adapter layer
First step of #650 (validator-agnostic core). Behavior-preserving.

- Move all zod-specific schema introspection and validation into
  src/zod/adapter.ts, implementing the neutral ValidatorAdapter
  interface (field extraction + validation + rich error info).
- Land validator-neutral ResolvedFieldMeta/ExtractedFields shapes and
  the shared alias/negation resolution policy in src/adapter/, so a
  future valibot adapter reuses the same field-meta assembly.
- Describe built-in commands (completion, skills, __complete, the
  politty bin) with internal validator-free descriptors; replace the
  z.registry-based arg metadata store with a WeakMap.
- Validate SKILL.md frontmatter without zod; keep the deprecated
  skillFrontmatterSchema export for backwards compatibility.
- As a result, no politty entry point except politty/skill reaches zod
  at runtime; zod loads only when a user schema flows through parsing
  or validation.
