+++ src/completion/dynamic/complete-command.ts
@@ -13,8 +13,11 @@
  *   fish: value\tdescription pairs, last line :directive
  */
 
-import { z } from "zod";
-import { arg } from "../../core/arg-registry.js";
+import {
+  internalArgs,
+  internalField,
+  type InferInternalArgs,
+} from "../../adapter/internal-args.js";
 import { defineCommand } from "../../core/command.js";
 import type { AnyCommand, ArgsSchema, Command } from "../../types.js";
 import { detectInlineOptionPrefix, generateCandidates } from "./candidate-generator.js";
@@ -24,19 +27,18 @@ import { formatForShell } from "./shell-formatter.js";
 /**
  * Schema for the __complete command
  */
-const completeArgsSchema = z.object({
-  shell: arg(z.enum(["bash", "zsh", "fish"]), {
+const completeArgsSchema = internalArgs({
+  shell: internalField.enum(["bash", "zsh", "fish"], {
     description: "Target shell for output formatting",
   }),
   // The arguments to complete are passed after --
-  args: arg(z.array(z.string()).default([]), {
+  args: internalField.stringArray({
     positional: true,
     description: "Arguments to complete",
-    variadic: true,
   }),
 });
 
-type CompleteArgs = z.infer<typeof completeArgsSchema>;
+type CompleteArgs = InferInternalArgs<typeof completeArgsSchema>;
 
 /**
  * Create the dynamic completion command
