+++ src/completion/index.ts
@@ -24,8 +24,7 @@
  * ```
  */
 
-import { z } from "zod";
-import { arg } from "../core/arg-registry.js";
+import { internalArgs, internalField, type InferInternalArgs } from "../adapter/internal-args.js";
 import { defineCommand } from "../core/command.js";
 import type { AnyCommand, ArgsSchema, Command } from "../types.js";
 import { generateBashCompletion } from "./bash.js";
@@ -168,73 +167,67 @@ export function detectShell(): ShellType | null {
 /**
  * Schema for the completion command arguments
  */
-const completionArgsSchema = z.object({
-  shell: arg(
-    z
-      .enum(["bash", "zsh", "fish"])
-      .optional()
-      .describe("Shell type (auto-detected if not specified)"),
-    {
-      positional: true,
-      description: "Shell type (bash, zsh, or fish)",
-      placeholder: "SHELL",
-    },
-  ),
-  instructions: arg(z.boolean().default(false), {
+const completionArgsSchema = internalArgs({
+  shell: internalField.optionalEnum(["bash", "zsh", "fish"], {
+    positional: true,
+    description: "Shell type (bash, zsh, or fish)",
+    placeholder: "SHELL",
+  }),
+  instructions: internalField.boolean({
     alias: "i",
     description: "Show installation instructions",
   }),
-  loader: arg(z.boolean().default(false), {
+  loader: internalField.boolean({
     description:
       "Print just the rc loader snippet (bash/zsh). Add it to ~/.bashrc or ~/.zshrc; it auto-regenerates the cache when the binary changes.",
   }),
-  install: arg(z.boolean().default(false), {
+  install: internalField.boolean({
     description:
       "Write the completion script to its on-disk cache (bash/zsh) or autoload location (fish) instead of printing it.",
   }),
-  static: arg(z.boolean().default(false), {
+  static: internalField.boolean({
     description: "Generate the legacy static completion script with command metadata baked in.",
   }),
-  dispatcher: arg(z.boolean().default(false), {
+  dispatcher: internalField.boolean({
     description: "Generate the runtime dispatcher completion script. This is the default.",
   }),
-  worker: arg(z.boolean().default(false), {
+  worker: internalField.boolean({
     description: "Generate an internal static worker artifact for dispatcher mode.",
   }),
 });
 
-type CompletionArgs = z.infer<typeof completionArgsSchema>;
+type CompletionArgs = InferInternalArgs<typeof completionArgsSchema>;
 
-const refreshArgsSchema = z.object({
-  shell: arg(z.enum(["bash", "zsh", "fish"]), {
+const refreshArgsSchema = internalArgs({
+  shell: internalField.enum(["bash", "zsh", "fish"], {
     positional: true,
     description: "Shell to refresh",
     placeholder: "SHELL",
   }),
-  target: arg(z.string().optional(), {
+  target: internalField.optionalString({
     positional: true,
     description: "Existing politty-generated completion file to refresh",
     placeholder: "TARGET",
   }),
-  static: arg(z.boolean().default(false), {
+  static: internalField.boolean({
     description: "Refresh using the legacy static completion script mode.",
   }),
-  worker: arg(z.boolean().default(false), {
+  worker: internalField.boolean({
     description: "Refresh an internal static worker completion script.",
   }),
 });
 
-type RefreshArgs = z.infer<typeof refreshArgsSchema>;
+type RefreshArgs = InferInternalArgs<typeof refreshArgsSchema>;
 
-const workerPathArgsSchema = z.object({
-  shell: arg(z.enum(["bash", "zsh", "fish"]), {
+const workerPathArgsSchema = internalArgs({
+  shell: internalField.enum(["bash", "zsh", "fish"], {
     positional: true,
     description: "Shell worker to locate",
     placeholder: "SHELL",
   }),
 });
 
-type WorkerPathArgs = z.infer<typeof workerPathArgsSchema>;
+type WorkerPathArgs = InferInternalArgs<typeof workerPathArgsSchema>;
 
 /**
  * Create a completion subcommand for your CLI
