+++ src/skill/frontmatter.ts
@@ -1,5 +1,5 @@
 import { parse as parseYaml } from "yaml";
-import { z } from "zod";
+import type { SkillFrontmatter } from "./types.js";
 
 /**
  * Skill name pattern from the Agent Skills specification:
@@ -19,40 +19,108 @@ const DESCRIPTION_MAX = 1024;
 const COMPATIBILITY_MAX = 500;
 
 /**
- * Zod schema for SKILL.md frontmatter.
+ * A single frontmatter validation problem, in the same `path`/`message`
+ * shape the scanner has always rendered into its diagnostics.
+ */
+export interface FrontmatterIssue {
+  path: string[];
+  message: string;
+}
+
+/**
+ * Validate parsed SKILL.md frontmatter against the Agent Skills
+ * specification (https://agentskills.io/specification).
  *
- * Strictly validates the fields defined in the Agent Skills specification
- * (https://agentskills.io/specification). Unknown fields are preserved via
- * `.passthrough()` so spec extensions and vendor keys round-trip intact.
+ * Validates the spec-defined fields strictly; unknown fields are preserved
+ * (passthrough) so spec extensions and vendor keys round-trip intact.
+ * Mirrors the deprecated `skillFrontmatterSchema` zod export
+ * (frontmatter-schema.ts) — keep the two in sync when the spec changes.
  *
  * Provenance / ownership for politty-managed installs is recorded under
  * `metadata["politty-cli"]` as `"{packageName}:{cliName}"`.
  */
-export const skillFrontmatterSchema = z
-  .object({
-    /** Skill identifier. Lowercase alphanumerics + hyphens, 1..64 chars. */
-    name: z.string().min(1).max(NAME_MAX).regex(SKILL_NAME_PATTERN, {
+export function validateSkillFrontmatter(
+  data: Record<string, unknown>,
+): { success: true; data: SkillFrontmatter } | { success: false; issues: FrontmatterIssue[] } {
+  const issues: FrontmatterIssue[] = [];
+
+  const checkString = (
+    key: string,
+    opts: { required?: boolean; min?: number; max?: number },
+  ): void => {
+    const value = data[key];
+    if (value === undefined) {
+      if (opts.required) {
+        issues.push({ path: [key], message: "Invalid input: expected string, received undefined" });
+      }
+      return;
+    }
+    if (typeof value !== "string") {
+      issues.push({
+        path: [key],
+        message: `Invalid input: expected string, received ${typeof value}`,
+      });
+      return;
+    }
+    if (opts.min !== undefined && value.length < opts.min) {
+      issues.push({
+        path: [key],
+        message: `Too small: expected string to have >=${opts.min} characters`,
+      });
+    }
+    if (opts.max !== undefined && value.length > opts.max) {
+      issues.push({
+        path: [key],
+        message: `Too big: expected string to have <=${opts.max} characters`,
+      });
+    }
+  };
+
+  checkString("name", { required: true, min: 1, max: NAME_MAX });
+  if (typeof data.name === "string" && !SKILL_NAME_PATTERN.test(data.name)) {
+    issues.push({
+      path: ["name"],
       message: "name must be lowercase alphanumerics separated by single hyphens",
-    }),
-    /** Human-readable description (1..1024 chars). */
-    description: z.string().min(1).max(DESCRIPTION_MAX),
-    /** SPDX license identifier or free-form string. */
-    license: z.string().min(1).optional(),
-    /** Runtime / tool compatibility string (<=500 chars). */
-    compatibility: z.string().max(COMPATIBILITY_MAX).optional(),
-    /** Metadata map (spec: string keys, string values). */
-    metadata: z.record(z.string(), z.string()).optional(),
-    /** Experimental spec field. */
-    "allowed-tools": z.string().optional(),
-  })
-  .passthrough();
+    });
+  }
+  checkString("description", { required: true, min: 1, max: DESCRIPTION_MAX });
+  checkString("license", { min: 1 });
+  checkString("compatibility", { max: COMPATIBILITY_MAX });
+  checkString("allowed-tools", {});
+
+  const metadata = data.metadata;
+  if (metadata !== undefined) {
+    if (typeof metadata !== "object" || metadata === null || Array.isArray(metadata)) {
+      issues.push({
+        path: ["metadata"],
+        message: `Invalid input: expected record, received ${metadata === null ? "null" : Array.isArray(metadata) ? "array" : typeof metadata}`,
+      });
+    } else {
+      for (const [key, value] of Object.entries(metadata)) {
+        if (typeof value !== "string") {
+          issues.push({
+            path: ["metadata", key],
+            message: `Invalid input: expected string, received ${typeof value}`,
+          });
+        }
+      }
+    }
+  }
+
+  if (issues.length > 0) {
+    return { success: false, issues };
+  }
+  // Unknown keys pass through untouched, matching the historical
+  // z.object(...).passthrough() behavior.
+  return { success: true, data: { ...data } as SkillFrontmatter };
+}
 
 /**
  * Result of parsing a SKILL.md file.
  */
 export interface ParsedSkillMd {
   /** Parsed and validated frontmatter */
-  frontmatter: z.infer<typeof skillFrontmatterSchema>;
+  frontmatter: SkillFrontmatter;
   /** Markdown body (content after frontmatter) */
   body: string;
   /** Full raw content */
@@ -72,7 +140,7 @@ const FRONTMATTER_PATTERN = /^\uFEFF?---[ \t]*\r?\n([\s\S]*?)\r?\n---[ \t]*(?:\r
  * inside failed to parse, so the scanner can distinguish "invalid YAML"
  * from "missing required field" in its diagnostics. A non-object root
  * (e.g. a top-level YAML list) also returns empty `data` without
- * `parseError` — Zod's schema validation surfaces that case clearly
+ * `parseError` — schema validation surfaces that case clearly
  * enough on its own.
  *
  * @example
@@ -117,7 +185,7 @@ export function parseFrontmatter(content: string): {
 /**
  * Root-level plain-object check. Rejects Dates, Maps, and custom tagged types
  * at the root of the parsed YAML; nested values are still validated by the
- * Zod schema that consumes this data.
+ * frontmatter validation that consumes this data.
  */
 function isPlainObject(value: unknown): value is Record<string, unknown> {
   if (value == null || typeof value !== "object" || Array.isArray(value)) return false;
@@ -133,7 +201,7 @@ function isPlainObject(value: unknown): value is Record<string, unknown> {
  */
 export function parseSkillMd(content: string): ParsedSkillMd | null {
   const { data, body } = parseFrontmatter(content);
-  const result = skillFrontmatterSchema.safeParse(data);
+  const result = validateSkillFrontmatter(data);
 
   if (!result.success) {
     return null;
