+++ src/validator/error-formatter.ts
@@ -1,5 +1,5 @@
+import type { ValidationError } from "../adapter/types.js";
 import { styles, symbols } from "../output/logger.js";
-import type { ValidationError } from "./zod-validator.js";
 
 /**
  * Calculate Levenshtein distance between two strings
